Update Dataset and Model
Eight classes were set. 0 for Cl and Br or more than four Cl or more than two Br; 1 for one Br or three Cl; 2 for one or two Cl; 3 for element with CHONPFIHoNaAlAs; 4 for hydro isomers;  5 for Fe; 6 for B; 7 for Se; 8 for not more than four S; 9 for not less than four S;....13.

Model was changed to only one nn model.